fix: Hide settings files from recents
We should hide those files in recents since we already hide theme in the
other screens.

I first tried to filter out settings files in the dataproxy request, but
files don't have paths in couchdb ou pouchdb databases. Looks like it will
not be a big deal to have a little bit less thant 50 files in the
results.